Naming: replace "timeout" and "expire" and "expiration" with EoL or End-of-Life.
This commit tightens up the verbiage used all through-out the class for pointing to and talking about the moment when a dismissal becomes undismissed.

After more synonym searching than I'd lie to admit, and after deciding against having Dawn of the Dead references permanently ratified into the API, I've settled on "eol" for the suffix of the user-meta lifespan key.

🧟